chore: moves database list into database wrapper
