--- app/components/database-list/types.d.ts
@@ -1,7 +1,0 @@
-import { DatabaseGame } from "../types"
-
-export type DatabaseListProps = {
-  games: DatabaseGame[]
-  label: string
-  id: string
-}
