+++ app/components/database-wrapper/ListToolbar.tsx
@@ -1,6 +1,15 @@
 "use client"
 
-import { ListToolbarProps } from "./types"
+import { Dispatch, SetStateAction } from "react"
+
+export type ListToolbarProps = {
+  label: string
+  pluralLabel: string
+  itemsLength: number
+  id: string
+  opened: boolean
+  setOpened: Dispatch<SetStateAction<boolean>>
+}
 
 export function ListToolbar({
   itemsLength,
