--- app/components/list-toolbar/types.d.ts
@@ -1,10 +1,0 @@
-import { Dispatch, SetStateAction } from "react"
-
-export type ListToolbarProps = {
-  label: string
-  pluralLabel: string
-  itemsLength: number
-  id: string
-  opened: boolean
-  setOpened: Dispatch<SetStateAction<boolean>>
-}
